made it actually show up

+++ app/src/main/java/com/example/earthquake/Geometry.kt
@@ -4,5 +4,5 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class Geometry(
-    val coordinates : List<Int>
+    val coordinates : List<Double>
 ) : Parcelable
